--- src/libraries/ApiAccountGet.ts
@@ -1,1 +1,0 @@
-export type ApiAccountGet = <T>(endpoint: string) => Promise<T>;
