--- src/libraries/ApiGet.ts
@@ -1,1 +1,0 @@
-export type ApiGet = <T>(endpoint: string) => Promise<T>;
