--- src/libraries/ApiPost.ts
@@ -1,3 +1,0 @@
-export type ApiPost = <T, P = any>(
-  endpoint: string
-) => <D = P>(postData: D) => Promise<T>;
