--- src/libraries/AxiosClient.ts
@@ -1,5 +1,0 @@
-import { AxiosRequestConfig, AxiosResponse } from 'axios';
-
-export type AxiosClient<T> = (
-  axiosConfig: AxiosRequestConfig
-) => Promise<AxiosResponse<T>>;
