--- src/libraries/RawApiAccountGet.ts
@@ -1,7 +1,0 @@
-import { ApiGet, Credentials } from '.';
-
-// - Copyright (c) Benjamin Vincent Kasapoglu (Luxcium). All rights reserved.
-// - Licensed under the MIT License.
-// - See License.txt in the project root for license information.
-
-export type RawApiAccountGet = (crendentials: Credentials) => ApiGet;
