--- src/libraries/RawApiGet.ts
@@ -1,2 +1,0 @@
-import { ApiGet, Credentials } from './';
-export type RawApiGet = (crendentials: Credentials) => ApiGet;
