--- src/libraries/RawApiPost.ts
@@ -1,2 +1,0 @@
-import { ApiPost, Credentials } from './';
-export type RawApiPost = (crendentials: Credentials) => ApiPost;
